Refactored tests to use helper functions
